extensions: inline S3ProvisionPipeline
