+++ extensions/aws/s3/provision/src/main/java/org/eclipse/dataspaceconnector/provision/aws/s3/S3BucketProvisioner.java
@@ -16,6 +16,7 @@ package org.eclipse.dataspaceconnector.provision.aws.s3;
 
 import net.jodah.failsafe.Failsafe;
 import net.jodah.failsafe.RetryPolicy;
+import org.eclipse.dataspaceconnector.provision.aws.AwsTemporarySecretToken;
 import org.eclipse.dataspaceconnector.provision.aws.provider.ClientProvider;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.eclipse.dataspaceconnector.spi.transfer.provision.DeprovisionResponse;
@@ -23,12 +24,21 @@ import org.eclipse.dataspaceconnector.spi.transfer.provision.ProvisionResponse;
 import org.eclipse.dataspaceconnector.spi.transfer.provision.Provisioner;
 import org.eclipse.dataspaceconnector.spi.types.domain.transfer.ProvisionedResource;
 import org.eclipse.dataspaceconnector.spi.types.domain.transfer.ResourceDefinition;
+import software.amazon.awssdk.awscore.exception.AwsServiceException;
 import software.amazon.awssdk.services.iam.IamAsyncClient;
+import software.amazon.awssdk.services.iam.model.CreateRoleRequest;
+import software.amazon.awssdk.services.iam.model.CreateRoleResponse;
 import software.amazon.awssdk.services.iam.model.DeleteRolePolicyRequest;
 import software.amazon.awssdk.services.iam.model.DeleteRolePolicyResponse;
 import software.amazon.awssdk.services.iam.model.DeleteRoleRequest;
 import software.amazon.awssdk.services.iam.model.DeleteRoleResponse;
+import software.amazon.awssdk.services.iam.model.GetUserResponse;
+import software.amazon.awssdk.services.iam.model.PutRolePolicyRequest;
+import software.amazon.awssdk.services.iam.model.Role;
+import software.amazon.awssdk.services.iam.model.Tag;
 import software.amazon.awssdk.services.s3.S3AsyncClient;
+import software.amazon.awssdk.services.s3.model.CreateBucketConfiguration;
+import software.amazon.awssdk.services.s3.model.CreateBucketRequest;
 import software.amazon.awssdk.services.s3.model.Delete;
 import software.amazon.awssdk.services.s3.model.DeleteBucketRequest;
 import software.amazon.awssdk.services.s3.model.DeleteBucketResponse;
@@ -37,16 +47,48 @@ import software.amazon.awssdk.services.s3.model.DeleteObjectsResponse;
 import software.amazon.awssdk.services.s3.model.ListObjectsV2Request;
 import software.amazon.awssdk.services.s3.model.ListObjectsV2Response;
 import software.amazon.awssdk.services.s3.model.ObjectIdentifier;
+import software.amazon.awssdk.services.sts.StsAsyncClient;
+import software.amazon.awssdk.services.sts.model.AssumeRoleRequest;
+import software.amazon.awssdk.services.sts.model.AssumeRoleResponse;
+import software.amazon.awssdk.utils.Pair;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.stream.Collectors;
 
+import static java.lang.String.format;
 import static java.util.stream.Collectors.joining;
 
 /**
  * Asynchronously provisions S3 buckets.
  */
 public class S3BucketProvisioner implements Provisioner<S3BucketResourceDefinition, S3BucketProvisionedResource> {
+
+    // Do not modify this trust policy
+    private static final String ASSUME_ROLE_TRUST = "{" +
+            "  \"Version\": \"2012-10-17\"," +
+            "  \"Statement\": [" +
+            "    {" +
+            "      \"Effect\": \"Allow\"," +
+            "      \"Principal\": {" +
+            "        \"AWS\": \"%s\"" +
+            "      }," +
+            "      \"Action\": \"sts:AssumeRole\"" +
+            "    }" +
+            "  ]" +
+            "}";
+    // Do not modify this bucket policy
+    private static final String BUCKET_POLICY = "{" +
+            "    \"Version\": \"2012-10-17\"," +
+            "    \"Statement\": [" +
+            "        {" +
+            "            \"Sid\": \"TemporaryAccess\", " +
+            "            \"Effect\": \"Allow\"," +
+            "            \"Action\": \"s3:PutObject\"," +
+            "            \"Resource\": \"arn:aws:s3:::%s/*\"" +
+            "        }" +
+            "    ]" +
+            "}";
+
     private final ClientProvider clientProvider;
     private final int sessionDuration;
     private final Monitor monitor;
@@ -64,7 +106,9 @@ public class S3BucketProvisioner implements Provisioner<S3BucketResourceDefiniti
         this.clientProvider = clientProvider;
         this.sessionDuration = sessionDuration;
         this.monitor = monitor;
-        this.retryPolicy = retryPolicy;
+        this.retryPolicy = retryPolicy.copy()
+                .withMaxRetries(10)
+                .handle(AwsServiceException.class);;
     }
 
     @Override
@@ -79,14 +123,22 @@ public class S3BucketProvisioner implements Provisioner<S3BucketResourceDefiniti
 
     @Override
     public CompletableFuture<ProvisionResponse> provision(S3BucketResourceDefinition resourceDefinition) {
-        S3ProvisionPipeline pipeline = S3ProvisionPipeline.Builder.newInstance(retryPolicy)
-                .resourceDefinition(resourceDefinition)
-                .clientProvider(clientProvider)
-                .sessionDuration(sessionDuration)
-                .monitor(monitor)
+        var s3AsyncClient = clientProvider.clientFor(S3AsyncClient.class, resourceDefinition.getRegionId());
+        var iamClient = clientProvider.clientFor(IamAsyncClient.class, resourceDefinition.getRegionId());
+        var stsClient = clientProvider.clientFor(StsAsyncClient.class, resourceDefinition.getRegionId());
+
+        var request = CreateBucketRequest.builder()
+                .bucket(resourceDefinition.getBucketName())
+                .createBucketConfiguration(CreateBucketConfiguration.builder().build())
                 .build();
 
-        return pipeline.provision();
+        return s3AsyncClient.createBucket(request)
+                .thenCompose(r -> getUser(iamClient))
+                .thenCompose(response -> createRole(iamClient, resourceDefinition, response))
+                .thenCompose(response -> createRolePolicy(iamClient, resourceDefinition, response))
+                .thenCompose(role -> assumeRole(stsClient, role))
+                .thenApply(result -> provisionSuccedeed(resourceDefinition, result))
+                .exceptionally(throwable -> provisionFailed(resourceDefinition, throwable));
     }
 
     @Override
@@ -104,6 +156,90 @@ public class S3BucketProvisioner implements Provisioner<S3BucketResourceDefiniti
                 .thenCompose(listAttachedRolePoliciesResponse -> deleteRolePolicy(iamClient, role))
                 .thenCompose(deleteRolePolicyResponse -> deleteRole(iamClient, role))
                 .thenApply(response -> DeprovisionResponse.Builder.newInstance().ok().resource(resource).build());
+    }
+
+    private CompletableFuture<Pair<Role, AssumeRoleResponse>> assumeRole(StsAsyncClient stsClient, Role role) {
+        return Failsafe.with(retryPolicy).getStageAsync(() -> {
+            monitor.debug("S3ProvisionPipeline: attempting to assume the role");
+            AssumeRoleRequest roleRequest = AssumeRoleRequest.builder()
+                    .roleArn(role.arn())
+                    .roleSessionName("transfer")
+                    .externalId("123")
+                    .build();
+
+            return stsClient.assumeRole(roleRequest)
+                    .thenApply(response -> Pair.of(role, response));
+        });
+    }
+
+    private CompletableFuture<Role> createRolePolicy(IamAsyncClient iamAsyncClient, S3BucketResourceDefinition resourceDefinition, CreateRoleResponse response) {
+        return Failsafe.with(retryPolicy).getStageAsync(() -> {
+            Role role = response.role();
+            PutRolePolicyRequest policyRequest = PutRolePolicyRequest.builder()
+                    .policyName(resourceDefinition.getTransferProcessId())
+                    .roleName(role.roleName())
+                    .policyDocument(format(BUCKET_POLICY, resourceDefinition.getBucketName()))
+                    .build();
+
+            monitor.debug("S3ProvisionPipeline: attach bucket policy to role " + role.arn());
+            return iamAsyncClient.putRolePolicy(policyRequest)
+                    .thenApply(policyResponse -> role);
+        });
+    }
+
+    private CompletableFuture<CreateRoleResponse> createRole(IamAsyncClient iamClient, S3BucketResourceDefinition resourceDefinition, GetUserResponse response) {
+        return Failsafe.with(retryPolicy).getStageAsync(() -> {
+            String userArn = response.user().arn();
+            Tag tag = Tag.builder().key("dataspaceconnector:process").value(resourceDefinition.getTransferProcessId()).build();
+
+            monitor.debug("S3ProvisionPipeline: create role for user" + userArn);
+            CreateRoleRequest createRoleRequest = CreateRoleRequest.builder()
+                    .roleName(resourceDefinition.getTransferProcessId()).description("EDC transfer process role")
+                    .assumeRolePolicyDocument(format(ASSUME_ROLE_TRUST, userArn))
+                    .maxSessionDuration(sessionDuration)
+                    .tags(tag)
+                    .build();
+
+            return iamClient.createRole(createRoleRequest);
+        });
+    }
+
+    private CompletableFuture<GetUserResponse> getUser(IamAsyncClient iamAsyncClient) {
+        return Failsafe.with(retryPolicy).getStageAsync(() -> {
+            monitor.debug("S3ProvisionPipeline: get user");
+            return iamAsyncClient.getUser();
+        });
+    }
+
+    private ProvisionResponse provisionSuccedeed(S3BucketResourceDefinition resourceDefinition, Pair<Role, AssumeRoleResponse> result) {
+        monitor.debug("S3ProvisionPipeline: STS credentials obtained, continuing...");
+        var resource = S3BucketProvisionedResource.Builder.newInstance()
+                .id(resourceDefinition.getBucketName())
+                .resourceDefinitionId(resourceDefinition.getId())
+                .region(resourceDefinition.getRegionId())
+                .bucketName(resourceDefinition.getBucketName())
+                .role(result.left().roleName())
+                .transferProcessId(resourceDefinition.getTransferProcessId())
+                .build();
+
+        var credentials = result.right().credentials();
+        var secretToken = new AwsTemporarySecretToken(credentials.accessKeyId(), credentials.secretAccessKey(), credentials.sessionToken(), credentials.expiration().toEpochMilli());
+
+        monitor.debug("Bucket request submitted: " + resourceDefinition.getBucketName());
+        return ProvisionResponse.Builder.newInstance().resource(resource).secretToken(secretToken).build();
+    }
+
+    private ProvisionResponse provisionFailed(S3BucketResourceDefinition resourceDefinition, Throwable exception) {
+        var exceptionToLog = exception.getCause() != null ? exception.getCause() : exception;
+        S3BucketProvisionedResource erroredResource = S3BucketProvisionedResource.Builder.newInstance()
+                .id(resourceDefinition.getBucketName())
+                .transferProcessId(resourceDefinition.getTransferProcessId())
+                .resourceDefinitionId(resourceDefinition.getId())
+                .error(true)
+                .errorMessage(exceptionToLog.getMessage())
+                .build();
+
+        return ProvisionResponse.Builder.newInstance().resource(erroredResource).build();
     }
 
     private CompletableFuture<DeleteRoleResponse> deleteRole(IamAsyncClient iamClient, String role) {
